+++ projects/core/src/valueConverters.ts
@@ -121,7 +121,7 @@ export class ValueConverters {
     toJson: (value) => ValueConverters.Number.toDb(value),
     fromInput: (x, type) => {
       let r = +x
-      if (!x) return undefined
+      if (x === null || x === undefined) return undefined
       return r
     },
     toInput: (x, type) => {